Extract the resolver part of CubeDockerConfigurator, so that it can be used in DockerRequirement. Add DockerRequirement. Use DockerRequirement.
