+++ docker/docker/src/main/java/org/arquillian/cube/docker/impl/requirement/RequiresDocker.java
@@ -1,0 +1,14 @@
+package org.arquillian.cube.docker.impl.requirement;
+
+import org.arquillian.cube.spi.requirement.Requires;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE, ElementType.METHOD})
+@Requires(DockerRequirement.class)
+public @interface RequiresDocker {
+}
