+++ docker/ftest-docker-compose/pom.xml
@@ -36,6 +36,11 @@
         </dependency>
         <dependency>
             <groupId>org.arquillian.cube</groupId>
+            <artifactId>arquillian-cube-requirement</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>org.arquillian.cube</groupId>
             <artifactId>arquillian-cube-docker</artifactId>
             <scope>test</scope>
         </dependency>
