+++ docker/ftest-docker-compose/src/test/java/org/arquillian/cube/servlet/HelloWorldServletTest.java
@@ -17,6 +17,8 @@ import org.arquillian.cube.ChangeLog;
 import org.arquillian.cube.CubeController;
 import org.arquillian.cube.CubeID;
 import org.arquillian.cube.TopContainer;
+import org.arquillian.cube.docker.impl.requirement.RequiresDocker;
+import org.arquillian.cube.requirement.ArquillianConditionalRunner;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
@@ -30,7 +32,8 @@ import org.junit.runner.RunWith;
 
 import com.github.dockerjava.api.DockerClient;
 
-@RunWith(Arquillian.class)
+@RequiresDocker
+@RunWith(ArquillianConditionalRunner.class)
 public class HelloWorldServletTest {
 
     @Deployment(testable=false)
@@ -43,7 +46,6 @@ public class HelloWorldServletTest {
 
     @Test
     public void should_parse_and_load_configuration_file(@ArquillianResource URL base) throws IOException {
-        
         URL obj = new URL(base, "HelloWorld");
         HttpURLConnection con = (HttpURLConnection) obj.openConnection();
         con.setRequestMethod("GET");
